[WIP][T00470]automotive-service-management: automotive-service-management

+++ automotive_service_management/wizard/cancel_service_appointment.py
@@ -1,11 +1,15 @@
 from odoo import models, fields, api
+from odoo.exceptions import ValidationError
 
 
-class CancelAppointment(models.Model):
+class CancelAppointment(models.TransientModel):
     _name = 'cancel.appointment'
     _description = 'cancel.appointment'
 
     appointment_ids = fields.Many2one(string="Select Appointment", comodel_name="service.appointment")
 
     def action_cancel_service_appointment(self):
-        self.appointment_ids.unlink()
+        if self.appointment_ids.state == 'appointed':
+            self.appointment_ids.unlink()
+        else:
+            raise ValidationError(f"In {self.appointment_ids.state} State You can Not Cancel Appointment")
